examples compile and run
