+++ examples/spherical_expansion_example.cc
@@ -30,9 +30,9 @@
 #include "structure_managers/adaptor_neighbour_list.hh"
 #include "structure_managers/make_structure_manager.hh"
 #include "rascal_utility.hh"
-#include "representations/representation_manager_sorted_coulomb.hh"
-#include "representations/representation_manager_spherical_expansion.hh"
-#include "representations/feature_manager_dense.hh"
+#include "representations/calculator_sorted_coulomb.hh"
+#include "representations/calculator_spherical_expansion.hh"
+#include "representations/calculator_spherical_invariants.hh"
 #include "basic_types.hh"
 #include "atomic_structure.hh"
 
@@ -48,10 +48,10 @@
 // using namespace std;
 using namespace rascal;  // NOLINT
 
-// using Representation_t = RepresentationManagerSOAP<
-// AdaptorStrict<AdaptorNeighbourList<StructureManagerCenters>>>;
-using Representation_t = RepresentationManagerSphericalExpansion<
-    AdaptorStrict<AdaptorNeighbourList<StructureManagerCenters>>>;
+using Representation_t = CalculatorSphericalExpansion;
+using Manager_t = AdaptorStrict<AdaptorNeighbourList<StructureManagerCenters>>;
+using Prop_t = typename CalculatorSphericalInvariants::Property_t<Manager_t>;
+using PropDer_t = typename CalculatorSphericalInvariants::PropertyGradient_t<Manager_t>;
 
 int main(int argc, char * argv[]) {
   if (argc < 2) {
@@ -92,8 +92,8 @@ int main(int argc, char * argv[]) {
                                    AdaptorNeighbourList, AdaptorStrict>(
           structure, adaptors);
 
-  Representation_t representation{manager, hypers};
-  representation.compute();
+  Representation_t representation{hypers};
+  representation.compute(manager);
 
   constexpr size_t n_centers_print{4};
   constexpr size_t n_neigh_print{1};
@@ -113,24 +113,29 @@ int main(int argc, char * argv[]) {
   std::cout << "Gradients are printed with: First Cartesian component, "
                "then species, along the rows; n-l-m along the columns.";
   std::cout << std::endl;
+
+  auto&& expansions_coefficients{manager->template get_property_ref<Prop_t>(representation.get_name())};
+  auto&& expansions_coefficients_gradient{manager->template get_property_ref<PropDer_t>(representation.get_gradient_name())};
+
+
   size_t center_count{0};
   for (auto center : manager) {
     if (center_count >= n_centers_print) {
       break;
     }
     size_t n_species_center{
-        representation.expansions_coefficients.get_keys(center).size()};
+        expansions_coefficients.get_keys(center).size()};
     std::cout << "============================" << std::endl;
     std::cout << "Center " << center.get_index();
     std::cout << " of type " << center.get_atom_type() << std::endl;
-    std::cout << representation.expansions_coefficients.get_dense_row(center);
+    std::cout << expansions_coefficients.get_dense_row(center);
     std::cout << std::endl;
     std::cout << "Gradient of this expansion wrt center pos: " << std::endl;
     std::cout << Eigen::Map<Eigen::MatrixXd>(
-        representation.expansions_coefficients_gradient.get_dense_row(center)
+        expansions_coefficients_gradient.get_dense_row(center)
             .data(),
         3 * n_species_center,
-        representation.expansions_coefficients_gradient.get_nb_comp());
+        expansions_coefficients_gradient.get_nb_comp());
     std::cout << std::endl;
     size_t neigh_count{0};
     for (auto neigh : center) {
@@ -138,7 +143,7 @@ int main(int argc, char * argv[]) {
         break;
       }
       auto keys_neigh =
-          representation.expansions_coefficients_gradient[neigh].get_keys();
+          expansions_coefficients_gradient[neigh].get_keys();
       std::cout << "Neighbour keys: ";
       for (auto key : keys_neigh) {
         std::cout << "(";
@@ -151,10 +156,10 @@ int main(int argc, char * argv[]) {
       std::cout << "Gradient of the above wrt atom " << neigh.back();
       std::cout << " of type " << neigh.get_atom_type() << std::endl;
       std::cout << Eigen::Map<Eigen::MatrixXd>(
-          representation.expansions_coefficients_gradient.get_dense_row(neigh)
+          expansions_coefficients_gradient.get_dense_row(neigh)
               .data(),
           3 * n_species_center,
-          representation.expansions_coefficients_gradient.get_nb_comp());
+          expansions_coefficients_gradient.get_nb_comp());
       std::cout << std::endl;
       ++neigh_count;
     }
